refactor(memory,node): single-source DoS caps + error taxonomy; extract domain model
Architecture touch-ups across the memory crate and its Node binding. No public
API or behavior change — all MCP error-mapping tests stay green:

- New `velesdb_memory::limits` (non-feature-gated) owns the DoS caps
  (MAX_FACT_BYTES / MAX_RECALL_LIMIT / MAX_WHY_HOPS) and the clamp helpers, so
  the MCP server and the Node `guards` stop re-declaring the same constants with
  a "keep in sync" comment — single source of truth, no drift.
- New `MemoryError::category() -> ErrorCategory {InvalidInput, NotFound,
  Internal}`. The MCP `to_error` and the Node `to_napi_err` now map the
  *category*, not each variant, so the client-facing taxonomy is identical
  across transports (and the JSON-RPC NotFound→invalid_params choice is explicit).
- Extract the domain data model (Link, Recollection, ColumnOp, ColumnFilter,
  MemoryNode, MemoryEdge, Explanation) into `model.rs`, separating the value
  types the layer exchanges from the service that computes them
  (service.rs 724 -> ~600 lines). Re-exports at the crate root are unchanged.
