+++ crates/velesdb-memory/src/lib.rs
@@ -25,11 +25,18 @@ pub mod embedder;
 pub mod error;
 pub mod extract;
 pub mod id;
+/// Resource caps (DoS limits) shared by every adapter — the single source of
+/// truth for fact size, recall limit, and `why` hop depth.
+pub mod limits;
 /// The MCP server transport. Gated behind the default `mcp` feature so library
 /// consumers (e.g. the language bindings) can depend on the memory core without
 /// pulling the `rmcp`/`tokio` server stack.
 #[cfg(feature = "mcp")]
 pub mod mcp;
+/// The domain data model — the value types the memory layer exchanges
+/// (`Link`, `Recollection`, `ColumnFilter`, `Explanation`, …), separate from the
+/// service that computes them.
+pub mod model;
 pub mod service;
 
 /// Default embedding dimension — the single source of truth, taken from the
@@ -39,13 +46,11 @@ pub const DEFAULT_DIMENSION: usize = velesdb_core::agent::DEFAULT_DIMENSION;
 pub use embedder::{DynEmbedder, EmbedError, Embedder, HashEmbedder};
 #[cfg(feature = "ollama")]
 pub use embedder::{OllamaEmbedder, DEFAULT_OLLAMA_MODEL, DEFAULT_OLLAMA_URL};
-pub use error::MemoryError;
+pub use error::{ErrorCategory, MemoryError};
 #[cfg(feature = "extract")]
 pub use extract::OllamaExtractor;
 pub use extract::{DynExtractor, ExtractError, ExtractedFact, Extractor};
 #[cfg(feature = "mcp")]
 pub use mcp::McpServer;
-pub use service::{
-    ColumnFilter, ColumnOp, Explanation, Link, MemoryEdge, MemoryNode, MemoryService, Metadata,
-    Recollection,
-};
+pub use model::{ColumnFilter, ColumnOp, Explanation, Link, MemoryEdge, MemoryNode, Recollection};
+pub use service::{MemoryService, Metadata};
